Added back updating feature for Lay
The previous feature was removed because insted of cloning Lay it restores it, however it doesnt work that way.  It now removes any ~/lay, runs 'lay-uninst', clones the repo again, moves it to /usr/local/lib, and runs install.sh (or alpine.sh if on alpine) in the folder.

Added: lay-update lay-update-alpine | Updated: install.sh alpine.sh